feat: moved subscriptions out of extension.ts
